requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
motor==3.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
faker==23.1.0
httpx==0.26.0
transformers==4.37.2
torch==2.2.0
chromadb==0.4.22
langchain-community==0.0.19
langchain==0.1.6
python-multipart==0.0.8
datasets==2.17.0
llama-cpp-python==0.2.39
pypdf==4.0.1
Hypercorn==0.16.0
